Refactored the file structure of the project
