+++ src/main/java/org/bitmaxsystems/oop2library/controllers/BasicUserDetailsController.java
@@ -10,12 +10,12 @@ import org.bitmaxsystems.oop2library.exceptions.DataValidationException;
 import org.bitmaxsystems.oop2library.models.dto.UserDataDTO;
 import org.bitmaxsystems.oop2library.models.users.User;
 import org.bitmaxsystems.oop2library.models.users.enums.UserRole;
-import org.bitmaxsystems.oop2library.util.service.DeleteUserService;
+import org.bitmaxsystems.oop2library.services.DeleteUserService;
 import org.bitmaxsystems.oop2library.util.UserManager;
-import org.bitmaxsystems.oop2library.util.contracts.IUserFormChain;
-import org.bitmaxsystems.oop2library.util.userformchain.UpdatePasswordChain;
-import org.bitmaxsystems.oop2library.util.userformchain.UpdateUserChain;
-import org.bitmaxsystems.oop2library.util.userformchain.VerifyUserDataChain;
+import org.bitmaxsystems.oop2library.util.chain.userform.contract.IUserFormChain;
+import org.bitmaxsystems.oop2library.util.chain.userform.UpdatePasswordChain;
+import org.bitmaxsystems.oop2library.util.chain.userform.UpdateUserChain;
+import org.bitmaxsystems.oop2library.util.chain.userform.VerifyUserDataChain;
 
 import java.util.Optional;
