+++ src/main/java/org/bitmaxsystems/oop2library/controllers/BookRegistryCreationController.java
@@ -8,9 +8,9 @@ import org.apache.logging.log4j.Logger;
 import org.bitmaxsystems.oop2library.exceptions.DataAlreadyExistException;
 import org.bitmaxsystems.oop2library.exceptions.DataValidationException;
 import org.bitmaxsystems.oop2library.models.dto.BookDataDTO;
-import org.bitmaxsystems.oop2library.util.bookformchain.CreateBookChain;
-import org.bitmaxsystems.oop2library.util.bookformchain.VerifyBookDataChain;
-import org.bitmaxsystems.oop2library.util.contracts.IBookFormChain;
+import org.bitmaxsystems.oop2library.util.chain.book.CreateBookChain;
+import org.bitmaxsystems.oop2library.util.chain.book.VerifyBookDataChain;
+import org.bitmaxsystems.oop2library.util.chain.book.contract.IBookFormChain;
 
 public class BookRegistryCreationController extends BaseBookRegistryFormController {
