+++ src/main/java/org/bitmaxsystems/oop2library/services/BookParameterService.java
@@ -1,4 +1,4 @@
-package org.bitmaxsystems.oop2library.util.service;
+package org.bitmaxsystems.oop2library.services;
 
 import org.bitmaxsystems.oop2library.exceptions.ChildRecordExistException;
 import org.bitmaxsystems.oop2library.exceptions.DataAlreadyExistException;
