+++ src/main/java/org/bitmaxsystems/oop2library/services/UserFormService.java
@@ -1,4 +1,4 @@
-package org.bitmaxsystems.oop2library.util.service;
+package org.bitmaxsystems.oop2library.services;
 
 import org.bitmaxsystems.oop2library.exceptions.FormAlreadyApprovedException;
 import org.bitmaxsystems.oop2library.models.form.UserForm;
