+++ src/test/java/BookParameterTest.java
@@ -8,7 +8,7 @@ import org.bitmaxsystems.oop2library.util.factory.AuthorFactory;
 import org.bitmaxsystems.oop2library.util.factory.GenreFactory;
 import org.bitmaxsystems.oop2library.util.factory.PublisherFactory;
 import org.bitmaxsystems.oop2library.util.factory.contract.BookParameterAbstractFactory;
-import org.bitmaxsystems.oop2library.util.service.BookParameterService;
+import org.bitmaxsystems.oop2library.services.BookParameterService;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
