+++ src/test/java/UserUpdateTest.java
@@ -1,12 +1,11 @@
-import org.bitmaxsystems.oop2library.config.HibernateInit;
 import org.bitmaxsystems.oop2library.models.dto.UserDataDTO;
 import org.bitmaxsystems.oop2library.models.users.User;
 import org.bitmaxsystems.oop2library.models.users.enums.UserRole;
 import org.bitmaxsystems.oop2library.repository.GenericRepository;
-import org.bitmaxsystems.oop2library.util.contracts.IUserFormChain;
-import org.bitmaxsystems.oop2library.util.userformchain.CreateUserChain;
-import org.bitmaxsystems.oop2library.util.userformchain.UpdatePasswordChain;
-import org.bitmaxsystems.oop2library.util.userformchain.UpdateUserChain;
+import org.bitmaxsystems.oop2library.util.chain.userform.contract.IUserFormChain;
+import org.bitmaxsystems.oop2library.util.chain.userform.CreateUserChain;
+import org.bitmaxsystems.oop2library.util.chain.userform.UpdatePasswordChain;
+import org.bitmaxsystems.oop2library.util.chain.userform.UpdateUserChain;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
